requirements/base.txt
asgiref
black
certifi
cffi
charset-normalizer
click
colorama
cryptography
distlib
dj-rest-auth
Django
django-allauth
djangorestframework
djangorestframework-simplejwt
drf-yasg
filelock
flake8
idna
inflection
iniconfig
isort
mccabe
mypy_extensions
oauthlib
packaging
pathspec
pipenv
platformdirs
pluggy
psycopg2-binary
pycodestyle
pycparser
pyflakes
Pygments
PyJWT
python-dotenv
python-json-logger
pytz
PyYAML
requests
requests-oauthlib
sqlparse
structlog
tzdata
uritemplate
urllib3
virtualenv
